added validation for form fields
