+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,8 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.FullScreen.Translucent">
+        android:theme="@style/Theme.FullScreen.Translucent"
+        >
         <activity android:name=".ui.SplashScreen"
             android:theme="@style/Theme.FullScreen">
             <intent-filter>
@@ -19,7 +20,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.MainActivity">
+        <activity android:name=".ui.MainActivity"
+            android:windowSoftInputMode="adjustResize|stateHidden"
+            >
 
         </activity>
     </application>
