+++ common/proguard-rules.pro
