+++ common/src/main/res/drawable-hdpi/unboarding_graphics.png
