+++ common/src/main/res/drawable-hdpi/wc_foundation.png
