+++ common/src/main/res/drawable-mdpi/unboarding_graphics.png
