+++ common/src/main/res/drawable-mdpi/wc_foundation.png
