+++ common/src/main/res/drawable-xhdpi/unboarding_graphics.png
