+++ common/src/main/res/drawable-xhdpi/wc_foundation.png
