+++ common/src/main/res/drawable-xxhdpi/unboarding_graphics.png
