+++ common/src/main/res/drawable/ic_baseline_mail_outline_24.xml
