+++ common/src/main/res/drawable/ic_baseline_phone_android_24.xml
