+++ common/src/main/res/values/colors.xml
