+++ settings.gradle
@@ -1,4 +1,4 @@
-include ':utils'
+include ':common'
 include ':navigation'
 include ':app'
 rootProject.name = "HAEMCAM"
